roachtest: port import/mixed-versions to the new framework
There are some minor differences:
- we no longer install fixtures (which weren't used anyway)
- we'll probably run the IMPORT multiple times throughout the test run.

Release note: None
