+++ pkg/cmd/roachtest/tests/BUILD.bazel
@@ -105,6 +105,7 @@ go_library(
         "mixed_version_change_replicas.go",
         "mixed_version_decl_schemachange_compat.go",
         "mixed_version_decommission.go",
+        "mixed_version_import.go",
         "mixed_version_job_compatibility_in_declarative_schema_changer.go",
         "mixed_version_schemachange.go",
         "multitenant.go",
